HR selection
This is version 0.5 - The conditional output does not work properly (only works for the first province of Alberta). However, when this is removed and the alternative dropdown menu is used - the data is filtered properly. I'm guessing it's something to do either with i) JS to grab the input$prov selection; or ii) or something missing on the UI side. The alternative code to access the less user friendly, but current only working option for the proper HR filtering of data, see the file titled `regular sidebar`, and copy and paste to replace the render sidebar code in the server side of the dashboard.