Add files via upload
Added an overall plot to visualize the increasing trend in Daily VOCs relative to Daily Cases. Made the Value Boxes reactive to Province input. Need to decide what to keep -- raw data and/ or averages. Also, B117 appears to be the dominant strain -- is it worth keeping distinctions between variants, and if so, for how long?